# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Happyfine Wholesalers API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filtering
      parameters:
        - name: category
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product with variants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/variants:
    get:
      operationId: listProductVariants
      tags: [products]
      summary: List variants for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
    post:
      operationId: createProductVariant
      tags: [products]
      summary: Create a variant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /variants/{id}:
    patch:
      operationId: updateVariant
      tags: [products]
      summary: Update a variant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
    delete:
      operationId: deleteVariant
      tags: [products]
      summary: Delete variant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CART ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, basePrice, categoryId, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        basePrice:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, draft]
        featured:
          type: boolean
        totalStock:
          type: integer
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductDetail:
      type: object
      required: [id, name, slug, basePrice, categoryId, status, variants]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        basePrice:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, draft]
        featured:
          type: boolean
        totalStock:
          type: integer
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, slug, basePrice, categoryId]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        basePrice:
          type: number
        compareAtPrice:
          type: number
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, draft]
        featured:
          type: boolean
    ProductPatch:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        basePrice:
          type: number
        compareAtPrice:
          type: number
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, draft]
        featured:
          type: boolean

    # ── VARIANT ──
    ProductVariant:
      type: object
      required: [id, productId, sku, price, stock]
      properties:
        id:
          type: integer
        productId:
          type: integer
        sku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        material:
          type: ["string", "null"]
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductVariantInput:
      type: object
      required: [sku, price, stock]
      properties:
        sku:
          type: string
          minLength: 1
        size:
          type: string
        color:
          type: string
        material:
          type: string
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: string
    ProductVariantPatch:
      type: object
      properties:
        sku:
          type: string
        size:
          type: string
        color:
          type: string
        material:
          type: string
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: string

    # ── CART ──
...