# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Happyfine Wholesalers API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  basePrice: number;
  /** @nullable */
  compareAtPrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  status: ProductStatus;
  featured?: boolean;
  totalStock?: number;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt?: string;
}