# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Happyfine Wholesalers API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from './productInputStatus';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  basePrice: number;
  compareAtPrice?: number;
  categoryId: number;
  imageUrl?: string;
  images?: string[];
  status?: ProductInputStatus;
  featured?: boolean;
}