# lib/api-zod/src/generated/types/productVariantPatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Happyfine Wholesalers API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductVariantPatch {
  sku?: string;
  size?: string;
  color?: string;
  material?: string;
  price?: number;
  stock?: number;
  imageUrl?: string;
}